Move keystore into own file
